requirements.txt
PyQt5
Pillow
keras
numpy